pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "wheel", "setuptools_scm>=8"]

[project]
authors = [{ name = "liu xue yan", email = "liu_xue_yan@foxmail.com" }]
description = "Extending PyYAML with a custom constructor for including YAML files within YAML files"
name = "pyyaml-include"
readme = "README.md"

requires-python = ">=3.8"

dependencies = ["PyYAML>=6.0,<7.0", "fsspec>=2021.04.0"]

keywords = ["yaml", "PyYAML", "include"]

license = { text = "GPLv3+" }

classifiers = [
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing :: Markup",
]

version = "2.0dev1"
# dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/tanbro/pyyaml-include"
Repository = "https://github.com/tanbro/pyyaml-include.git"
Documentation = "https://pyyaml-include.readthedocs.io/en/latest/"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/yamlinclude/_version.py"
